+++ doc/changeLog.md
@@ -1,3 +1,5 @@
+### 2018-05-18
+- DateRangePicker更新为antdesign
 ### 2018-05-09
 - 新增业务逻辑组件说明 Common
 ### 2018-04-28